+++ packages/sdk-akamai/demo/utils.js
@@ -61,6 +61,28 @@ const createClientRequest = (req, config) => {
       },
       implementationDetails: {
         name: "server-side",
+        version: "0.0.5",
+        environment: "server",
+      },
+    },
+    query: {
+      identity: {
+        fetch: ["ECID"],
+      },
+      personalization: {
+        schemas: [
+          "https://ns.adobe.com/personalization/default-content-item",
+          "https://ns.adobe.com/personalization/html-content-item",
+          "https://ns.adobe.com/personalization/json-content-item",
+          "https://ns.adobe.com/personalization/redirect-item",
+          "https://ns.adobe.com/personalization/dom-action",
+        ],
+        decisionScopes: [
+          "cdn-sdk-mbox-test",
+          "ss-control-test",
+          "previewDemo",
+          "mboxtestmay",
+        ],
       },
     },
     data: {
@@ -91,9 +113,11 @@ const getPersistedValues = (response) => {
   const locationHintIdHandle = response?.handle?.find(
     (payload) =>
       payload.type === "locationHint:result" &&
-      payload?.payload[0]?.scope === "EdgeNetwork",
+      payload?.payload.find((scope) => scope.scope === "EdgeNetwork")?.hint,
   );
-  const locationHintId = locationHintIdHandle?.payload[0]?.hint || "";
+  const locationHintId =
+    locationHintIdHandle?.payload.find((scope) => scope.scope === "EdgeNetwork")
+      ?.hint || "";
 
   return { ecid, locationHintId };
 };