+++ packages/sdk-nodejs/src/index.js
@@ -18,6 +18,7 @@ import {
 import { rng } from "./uuid/rng.js";
 import { httpRequestAdapterInstance } from "./httpRequestAdapter.js";
 import { md5 } from "./md5.js";
+import { createNodeScheduler } from "./scheduler.js";
 
 /**
  * The Client initialization method
@@ -29,5 +30,6 @@ export async function Client(clientOptions) {
   Container().registerInstance(TOKENS.HTTP_CLIENT, httpRequestAdapterInstance);
   Container().registerInstance(TOKENS.LOGGER, console);
   Container().registerInstance(TOKENS.MD5, md5);
+  Container().registerInstance(TOKENS.SCHEDULER, createNodeScheduler());
   return BaseClient(clientOptions);
 }